refactor: polish credit notification admin flows
